\name{SlideDesignerGUI-package}
\alias{SlideDesignerGUI-package}
\docType{package}
\title{Package for slide design layout}
\description{
  A package for recording layout of a slide design.
}
\details{
  \tabular{ll}{
    Package: \tab SlideDesignerGUI\cr
    Type: \tab Package\cr
    Version: \tab 1.3.1\cr
    Date: \tab 2010-10-08\cr
    License: \tab See LICENSE for license information\cr
  }

  For a complete list of functions, use \code{library(help="SlideDesignerGUI")}.
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SlideDesignerGUI"))}.
}
\author{P. Roebuck, \email{plroebuck@mdanderson.org}}
\keyword{package}

