\name{slideDesignerGUI}
\alias{slideDesignerGUI}
\alias{sdui}
\title{Graphical User Interface for slide design layout}
\description{
  Graphical User Interface for recording layout of a slide design.
}
\usage{
slideDesignerGUI()
sdui()
}
\details{
  The function \code{slideDesignerGUI} launches a Graphical User Interface
  for recording the layout of positive and negative controls of a slide
  design. The GUI uses Tk widgets (via the R TclTk interface by Peter Dalgaard)
  in order to provide a simple interface.

  The function \code{sdui} is an alias for the aforementioned function.
}
\author{
  P. Roebuck \email{plroebuck@mdanderson.org}
}
\examples{
  \dontrun{
slideDesignerGUI()
  }
}
\keyword{IO}

