% $Id$

\name{SlideDesignerGUI-package}
\alias{SlideDesignerGUI-package}
\docType{package}
\title{Package for slide design layout}
\description{
  A package for recording layout of a slide design.
}
\details{
  \tabular{ll}{
    Package: \tab SlideDesignerGUI\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("SlideDesignerGUI")$Version}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("SlideDesignerGUI")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("SlideDesignerGUI")$License}\cr
  }

  For a complete list of functions, use \code{library(help="SlideDesignerGUI")}.\cr
  For a high-level summary of the changes for each revision, use
    \code{file.show(system.file("NEWS", package="SlideDesignerGUI"))}.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

